+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/common/hardware/StandardTrackingWheelLocalizer.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.common;
+package org.firstinspires.ftc.teamcode.common.hardware;
 
 import androidx.annotation.NonNull;
